Add support for providing reference regs for Friendly Name Regs
Add support for providing a reference regs txt file to add Friendly Name
to output Regs. This only works with --final-provision.

An example file is providied.
Format to add a reg in the file is the following:
mmd_reg.reg Friendly Name

Option is enabled by default and takes the file reference_regs.txt as
source.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>